Handle the fact that nodes with eta_%d tags may not be actually the node you thought it was. For example Google Groups clones existing text elements in the page when the reply button is pressed. These elements have the attributes but crucially not listeners.
So now when we scan for text areas and we find one with an ETA we
check the list and see it really is the element we think it is. If not
we see if the gumdrop is nearby (and remove it) and re-do the text
area tagging.

The code is fairly clunky to do this, I'm sure there should be a
neater more Javascripty way....