From: "David J. Biesack" <David.Biesack@sas.com>
Here is a small patch that I have found convenient - it allows the
edit buffer to run in multibyte mode, so that it displays Unicode
characters correctly in Emacs 23. (I often edit wiki pages in which
others have inserted Unicode ellipses ... or left/right double quotes
or em dash. With emacs multibyte enabled, this characters display
correctly instead  of as binary bytes.