Incremental saves in text editor make updates to textarea.
This is accomplished by extending the communication protocol between
the extension and the edit server.  The edit server watches for
changes to the temp file it creates.  When a change is discovered, the
edit server responds as normal, except that it adds two new headers,
'x-file' and 'x-open'.  The value for 'x-file' is the name of the
temporary file.  The value for 'x-open' is set to 'true'.  When the
extension receives a response with 'x-open' set to 'true', it
immediately makes another request to the edit server, adding a header,
'x-file', whose value is the same as the value in the preceding
response.  If the edit server gets a request with a valid 'x-file'
header, it does not create a new temp file and spawn a new text editor
instance.  Instead, it watches the given file for changes.  The
process of the text editor is also monitored for termination.  A
terminated editor instance also triggers a response by the edit
server.

The Python edit server was the only one modified to handle the
changes.  It shouldn't be too difficult to update the Emacs-based edit
server in the future.