Use window-system to detect when we're on OS X
(featurep 'aquamacs) was false for me on the experimental Emacs Mac
port, causing edit-server to fail since make-frame-on-display seems to
only work when you're under X11.  I presume--but have not tested--that
it would also fail for the official Emacs NS port.

This new test on window-system should cover Aquamacs (tested), Emacs
NS port (untested), and the Emacs Mac port (tested).