Add textareas.css and a specific CSS ID for the edit buttons.
Reported as issue 41 the site:

http://www.thegatesnotes.com/Personal/Steve-Jobs

Has CSS which overode the size of all the image elements in a given
div box. As a result the button was stretched to fill the <div> area.
By specifying a specific CSS ID the size vould be forced back again.

I assume it's possible the CSS ID might clash with another sites but
hopefully that will be fairly rare.