Fix colour fading on <div> areas.
This seems to come down to the difference between rgb and rgba colours
which I get from the <textarea> and <div> css properties. As there
were bugs referring to this on the plugin page I updated to the latest
jquery.color.plugin code which also wanted a more up to date jQuery.