edit-server: send x-file and x-open headers for edit-server-save
This is part one of incremental save support. When we run C-x
C-s (edit-server-save) we add headers to the response. The browser
interprets those as an in-progress save and sends a fresh edit request
back.

At the moment this will create a brand new buffer. However once the
edit-server is aware of the x-file parameter we should be able to just
bring back the buffer we were editing.