Send textInput event after updating text
After updating the text send a `textInput` event appending a newline to
it. This will cause some sites (at least StackOverflow.com) to update
their preview of the rendered content.  Unfortunately this also causes
an extra newline to be added to the end of the area, but there tends to
be an extra one there anyway.