textarea.js: change the rules for ignoring hidden elements (#78)
On GitHub's Add SSH keys display the mutation event never picked up
the re-appearing hidden textarea form. The code was originally put in
to stop lots of random edit buttons appearing on Gmail due to it's
many hidden textareas. The change basically says we'll only ignore the
element if it is hidden itself and not due to it's parent. If it's
hidden due to it's parent then any tag we add will also be hidden.