Add initial support for "foreground" command
This is mainly for the benefit of Chromebooks who have a chroot'ed emacs that can run on the main display window. When enabled the /foreground command will make emacs pop up in the front of the display without an edit buffer when there is no text area in focus and the button is clicked.