xmlcomms.js: fixed up the foreground message handling
The previous method using the persistent tab connection has a few
problems. The principle failure was when there was no content script
to talk to. As the call is asynchronus we need to check for failure in
the callback which is only possible using sendMessage().

Also some clean-up of the tabs that seem to have sneaked into the code.