Fixup the img CSS in an attempt to fix #86
For starters the CSS wasn't being applied at all because I'd used an
id instead of class descriptor. Having done that I found the behaviour of
divs and textareas where different so I split the css rules into two.

I then experimented with Gmail's interface in an attempt to get the
edit button inside the sibling element. Unfortunately all the guides
I could find on CSS positioning said these things are done with
reference to the parent node, not the sibling. It might be for divs we
need to place the image inside the div to have enough control but this
then raises issues of how to clean up the contents before the
application does something with it. We after all are only meant to put
the users text in the box, not random html :-/

Also:

  * removed the tabs that snuck into that function