testareas.js: make it more obvious when hiding textareas (#93)
This is not a fix but will make it obvious in the console log when we
have hidden a text area. Unfortunately when the Blogger text area
comes into focus it doesn't seem to trigger a DOM update.

Maybe we should just skip adding buttons but still allow event mapping.