edit-server.el: do not call `set-buffer-multibyte` in the process buffer
In edit-server-accept, set-buffer-multibyte is called, but it's applied to
whatever buffer was previously active because the call is not wrapped with
`with-current-buffer`. This makes edit-server fail if that previously
active buffer is an indirect buffer (set-buffer-multibyte is illegal in an
indirect buffer).

Rather than add with-current-buffer, though, just remove the call
altogether. It is not necessary for the process buffer to be multibyte.