textareas: use a <div> with background-image for button (#120)
This work as the button is clickable but defining the location of the
icon in the CSS without hardcoding seems tricky. The image is now a
background image property of the div.

We also fix up the mutators to select against this and log removal of
images. At the same time we set the actual id of the button for easier
jquery selection.