edit-server: better document edit-server-new-frame-alist (#103/#109/#141)
Deciding when an where to create frames has proved to be a pain over
the years. As we now call make-frame directly explicitly mention how
window-system et all will affect that decision in the doctext. The
user has ultimate control by manipulating edit-server-new-frame-alist.
Otherwise emacs will open a frame on whatever is open currently, which
in emacs --daemon mode might not be really there.